Todo app

+++ routes/api.php
@@ -21,20 +21,24 @@ Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
+Route::middleware('auth:api')->get('/test', function() {
+    return "Test page";
+});
+
 // -- Category api urls
-Route::get('/categories', [CategoryController::class, 'getCategories']);
-Route::post('/categories', [CategoryController::class, 'createCategory']);
-Route::get('/categories/{categoryId}', [CategoryController::class, 'getCategory']);
-Route::patch('/categories/{categoryId}', [CategoryController::class, 'updateCategory']);
-Route::delete('/categories/{categoryId}', [CategoryController::class, 'deleteCategory']);
-Route::get('/categories/{categoryId}/products', [CategoryController::class, 'getProductsOfCategory']);
+Route::middleware('auth:api')->get('/categories', [CategoryController::class, 'getCategories']);
+Route::middleware('auth:api')->middleware('isAdmin')->post('/categories', [CategoryController::class, 'createCategory']);
+Route::middleware('auth:api')->get('/categories/{categoryId}', [CategoryController::class, 'getCategory']);
+Route::middleware('auth:api')->patch('/categories/{categoryId}', [CategoryController::class, 'updateCategory']);
+Route::middleware('auth:api')->delete('/categories/{categoryId}', [CategoryController::class, 'deleteCategory']);
+Route::middleware('auth:api')->get('/categories/{categoryId}/products', [CategoryController::class, 'getProductsOfCategory']);
 
 // -- Product api urls
-Route::get('/products', [ProductController::class, 'getProducts']);
-Route::post('/products', [ProductController::class, 'createProduct']);
-Route::get('/products/{productId}', [ProductController::class, 'getProduct']);
-Route::patch('/products/{productId}', [ProductController::class, 'updateProduct']);
-Route::delete('/products/{productId}', [ProductController::class, 'deleteProduct']);
+Route::middleware('auth:api')->get('/products', [ProductController::class, 'getProducts']);
+Route::middleware('auth:api')->post('/products', [ProductController::class, 'createProduct']);
+Route::middleware('auth:api')->get('/products/{productId}', [ProductController::class, 'getProduct']);
+Route::middleware('auth:api')->patch('/products/{productId}', [ProductController::class, 'updateProduct']);
+Route::middleware('auth:api')->delete('/products/{productId}', [ProductController::class, 'deleteProduct']);
 
 // -- Authentication api urls
 Route::post('/register', [AuthenticationController::class, 'register']);
